Release note content now support formio Form json instead of raw HTML
